AI_Project/requirements.txt
langchain>=0.0.375
openai>=1.0.0
google-api-python-client>=2.70.0
google-auth>=2.20.0
google-auth-httplib2>=0.1.0
python-dotenv>=1.0.0
requests>=2.28.0
APScheduler>=3.9.1
pytest>=7.0.0
email-validator>=1.3.1
